Små ændringer
Tilføjede delete confirmation og små UI ændringer

+++ pages/read.php
@@ -5,7 +5,6 @@
  * Date: 1/31/2019
  * Time: 9:21 AM
  */
-
 require("../extra/CRUD.php");
 //inkluder/kræv filer
 $title = "Produkter";
@@ -14,7 +13,8 @@ include("../extra/UI/header.php");
 $object = new CRUD();
 
 // Design table header skabelon
-$data = '<table width=\'80%\' border=1">
+$data = '<table class="table">
+             <thead>
                         <tr>
                             <th>Navn</th>
                             <th>Varenummer</th>
@@ -22,7 +22,7 @@ $data = '<table width=\'80%\' border=1">
                             <th>ID</th>
                             <th>Rediger</th>
                             <th>Slet</th>
-                        </tr>';
+                        </tr> </thead>';
 
 
 
@@ -32,7 +32,7 @@ $products = $object->Read();
 if (count($products) > 0) {
     foreach ($products as $product) {
 
-        $data .= '<tr bgcolor=\'#CCCCCC\'>
+        $data .= '<tbody> <tr class="active">
                 <td>' . $product['products_description_name'] . '</td>
                 <td>' . $product['products_reference'] . '</td>
                 <td>' . $product['products_price'] . '</td>
@@ -40,8 +40,8 @@ if (count($products) > 0) {
                 <td>
                 <a href="edit.php?id='.$product['products_id'].'" class="btn btn-warning">Rediger</a></td>
                 <td>
-                <a href="../extra/delete.php?id='.$product['products_id'].'" class="btn btn-danger">Slet</a></td>
-            </tr>';
+                <a href="../extra/delete.php?id='.$product['products_id'].'" class="btn btn-danger" onclick="return confirm(\'Er du sikker?\')">Slet</a></td>
+            </tr> </tbody>';
     }
 } else {
     // Data ikke fundet
@@ -50,11 +50,12 @@ if (count($products) > 0) {
 
 $data .= '</table>';
 
-echo $data;
+?> 
 
-?>
-<br>
-<br>
-<br>
-<a href="create.php?id=0">Tilføj produkt</a>
+    <?php echo $data; ?>   
+
+ <br>
+ <br>
+ <br>
+ <a class="btn btn-link" href="create.php?id=0">Tilføj produkt</a>
 <?php include("../extra/UI/footer.php"); ?>
